+++ frontend/src/views/routes/WorkoutLog.css
@@ -30,13 +30,14 @@
 
 .form-card h2 {
     font-size: 24px;
-    margin-bottom: 30px;
+    margin-bottom: 25px;
     color: #2C3E50;
     text-align: center;
 }
 
 .form-label {
     display: block;
+    text-align: left;
     margin-bottom: 5px;
     font-size: 14px;
     font-weight: bold;
@@ -79,7 +80,7 @@
     border-radius: 5px;
     font-size: 1rem;
     transition: border-color 0.3s;
-    margin-bottom: 10px;
+    margin-bottom: 6px;
 }
 
 .log-button {
